+++ doc/whats_new.rst
@@ -73,9 +73,13 @@ Changelog
 
     - Add interactive annotation mode to :meth:`mne.io.Raw.plot` (accessed by pressing 'a') by `Jaakko Leppakangas`_
 
+<<<<<<< 0c31b1e248bea43343bded9b7fa0e2d7aa664210
     - Add support for deleting all projectors or a list of indices in :meth:`mne.io.Raw.del_proj` by `Eric Larson`_
 
     - Add source space plotting with :meth:`mne.SourceSpaces.plot` using :func:`mne.viz.plot_trans` by `Eric Larson`_
+=======
+    - Add :func:`mne.decoding.get_coef` to retrieve and inverse the coefficients of a linear model, by `Jean-Remi King`_
+>>>>>>> ENH: get patterns coefs with example
 
     - Add support for reading in EGI MFF digitization coordinate files in :func:`mne.channels.read_dig_montage` by `Matt Boggess`_
 
@@ -172,6 +176,8 @@ API
     - MNE's additional files for the ``fsaverage`` head/brain model are now included in MNE-Python, and the now superfluous ``mne_root`` parameter to  :func:`create_default_subject` has been deprecated by `Christian Brodbeck`_
 
     - An ``overwrite=False`` default parameter has been added to :func:`write_source_spaces` to protect against accidental overwrites, by `Eric Larson`_
+
+    - The :class:`mne.decoding.LinearModel` class will no longer support `plot_filters` and `plot_patterns`, use mne.EvokedArray instead, by `Jean-Remi King`_
 
 .. _changes_0_13: